move Failsafe aspect to package
